update component layouts
