+++ client/src/components/app/app.scss
@@ -1,10 +1,12 @@
 
 .app-container{
-  position: relative;
+  display: flex;
+  justify-content: center;
+  position: absolute;
+  top: 100px;
   left: 100px;
-  height: calc(100vh - 100px);
-  width: calc(100vw - 100px)
-  
+  width: 100%;
+  height: 100%;
 }
 
 [data-initials]:before {
