+++ client/src/components/home/home.scss
@@ -1,6 +1,7 @@
 .home-container{
     display:flex;
     flex-direction: column;
+    width: 100%;
     
     h3{
         padding-left: 50px;
