+++ client/src/components/navigation/navBar.scss
@@ -1,15 +1,17 @@
 .side-nav{
     background: #181818;
     width: 100px;
-    height: 100%;
-    position: absolute;
+    height: 100vh;
+    position: -webkit-sticky;
+    position: sticky;
     top: 0;
     z-index: 2;
     min-height: 500px;
 
 
     ul{
-        position:absolute;
+        position: -webkit-sticky;
+        position: sticky;
         top: 120px;
         width: 100%;
         display: block;
