+++ client/src/components/recipes/addIngredient/addIngredient.scss
@@ -1,0 +1,4 @@
+.add-ingredient-container{
+    display: flex;
+    flex-direction: column;
+}
