+++ client/src/components/recipes/recipeCard/recipeCard.scss
@@ -3,6 +3,7 @@
     border: 1px solid #099bdd;
     border-radius: 3px;
     margin:3px;
+    width: 500px;
     .link{
         text-decoration: none;
     }
