+++ client/src/components/recipes/recipeCreate/recipeCreate.scss
@@ -1,0 +1,44 @@
+.create-new-recipe{
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    width: 100%;
+    min-width: 500px;
+    position: relative;
+    padding: 50px;
+    
+
+    .input{
+        display: flex;
+        align-items: flex-start;
+        margin: 10px;
+        input{
+            position: absolute;
+            left: 150px;
+            min-width: 200px;
+        }
+
+        
+    }
+    .instruction-container{
+        display: flex;
+        align-items: flex-start;
+        width: 100%;
+        height: 150px;
+        margin: 10px;
+        textarea{
+            position: absolute;
+            left: 150px;
+            min-width: 500px;
+            min-height: 100px;
+            resize: both;
+            
+        }
+    }
+
+    button{
+        position: relative;
+        
+        margin: 20px;
+    }
+}
