+++ client/src/components/recipes/recipeList/recipeList.scss
@@ -1,6 +1,7 @@
 .recipe-list-container{
     display:flex;
     flex-direction: column;
+    width: 100%;
     
     .top{
         display: flex;
