+++ client/src/components/users/chefList/chefList.scss
@@ -1,6 +1,7 @@
 .chef-list-container{
     display:flex;
     flex-direction: column;
+    width: 100%;
     
     .top{
         display: flex;
